show email when logged in
