+++ client/src/authentication/Register.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
-import { Link, Redirect } from "react-router-dom";
+import { Redirect } from "react-router-dom";
 import Cookies from 'js-cookie';
-
 import styled from 'styled-components';
 
 const Title = styled.h1`
@@ -13,10 +12,6 @@ const Title = styled.h1`
 const FormBox = styled.div`
   margin-top: 40px;
   text-align: center;
-`
-
-const LinkBox = styled.div`
-  margin-top: 20px;
 `
 
 const CenteredInput = styled.input`
@@ -32,7 +27,7 @@ class Login extends Component {
   }
 
   change = this.change.bind(this);
-  processLogin = this.processLogin.bind(this);
+  processRegistration = this.processRegistration.bind(this);
 
   change(e) {
     this.setState({
@@ -40,10 +35,10 @@ class Login extends Component {
     })
   }
 
-  processLogin(e) {
+  processRegistration(e) {
     e.preventDefault();
 
-    fetch('api/login.json', {
+    fetch('api/users.json', {
       method: 'post',
       credentials: 'same-origin',
       mode: 'same-origin',
@@ -61,6 +56,7 @@ class Login extends Component {
       if (data.auth_token) {
         Cookies.set("auth_token", data.auth_token);
         Cookies.set("user_id", data.user_id);
+        Cookies.set("email", data.email);
         this.setState({ logged_in: true })
       }
     }.bind(this)).catch(function(ex) {
@@ -72,40 +68,37 @@ class Login extends Component {
     return (
       <div>
         { (this.state.logged_in) ?
-          (
-            <Redirect to="/" />
-          ) : (
+        (
+          <Redirect to="/" />
+        ) : (
           <div>
-            <Title>Welcome Back</Title>
-            <div className="row justify-content-center">
-            <FormBox className="col-sm-6 col-lg4">
-              <form onSubmit={this.processLogin}>
-                <div className="form-group">
-                  <CenteredInput
-                    className="form-item form-control"
-                    placeholder="email..."
-                    name="email"
-                    type="text"
-                    onChange={this.change}
-                  />
-                </div>
-                <div className="form-group">
-                  <CenteredInput
-                    className="form-item form-control"
-                    placeholder="password..."
-                    name="password"
-                    type="password"
-                    onChange={this.change}
-                  />
-                </div>
-                <button type="submit" className="btn btn-primary">Login</button>
-              </form>
-              <LinkBox>
-                <Link to="/register">New Here? Register.</Link>
-              </LinkBox>
-            </FormBox>
-            </div>
+          <Title>Register to Enjoy Benefit</Title>
+          <div className="row justify-content-center">
+          <FormBox className="col-sm-6 col-lg4">
+            <form onSubmit={this.processRegistration}>
+              <div className="form-group">
+                <CenteredInput
+                  className="form-item form-control"
+                  placeholder="email..."
+                  name="email"
+                  type="text"
+                  onChange={this.change}
+                />
+              </div>
+              <div className="form-group">
+                <CenteredInput
+                  className="form-item form-control"
+                  placeholder="password..."
+                  name="password"
+                  type="password"
+                  onChange={this.change}
+                />
+              </div>
+              <button type="submit" className="btn btn-primary">Register</button>
+            </form>
+          </FormBox>
           </div>
+        </div>
         )}
       </div>
     );
